+++ kong/db/schema/plugin_loader.lua
@@ -11,6 +11,7 @@ local typedefs = require "kong.db.schema.typedefs"
 local Entity = require "kong.db.schema.entity"
 local utils = require "kong.tools.utils"
 local plugin_servers = require "kong.runloop.plugin_servers"
+local utils_toposort = utils.topological_sort
 
 
 local plugin_loader = {}
@@ -21,6 +22,54 @@ local next = next
 local type = type
 local insert = table.insert
 local ipairs = ipairs
+
+
+-- Given a hash of daos_schemas (a hash of tables,
+-- direct parsing of a plugin's daos.lua file) return an array
+-- of schemas in which:
+-- * If entity B has a foreign key to A, then B appears after A
+-- * If there's no foreign keys, schemas are sorted alphabetically by name
+local function sort_daos_schemas_topologically(daos_schemas)
+  local schema_defs = {}
+  local len = 0
+  local schema_defs_by_name = {}
+
+  for name, schema_def in pairs(daos_schemas) do
+    if name ~= "tables" or schema_def.fields then
+      len = len + 1
+      schema_defs[len] = schema_def
+      schema_defs_by_name[schema_def.name] = schema_def
+    end
+  end
+
+  -- initially sort by schema name
+  table.sort(schema_defs, function(a, b)
+    return a.name > b.name
+  end)
+
+  -- given a schema_def, return all the schema defs to which it has references
+  -- (and are on the list of schemas provided)
+  local get_schema_def_neighbors = function(schema_def)
+    local neighbors = {}
+    local neighbors_len = 0
+    local neighbor
+
+    for _, field in ipairs(schema_def.fields) do
+      if field.type == "foreign"  then
+        neighbor = schema_defs_by_name[field.reference] -- services
+        if neighbor then
+          neighbors_len = neighbors_len + 1
+          neighbors[neighbors_len] = neighbor
+        end
+        -- else the neighbor points to an unknown/uninteresting schema. This might happen in tests.
+      end
+    end
+
+    return neighbors
+  end
+
+  return utils_toposort(schema_defs, get_schema_def_neighbors)
+end
 
 
 --- Check if a string is a parseable URL.
@@ -251,16 +300,21 @@ function plugin_loader.load_entities(plugin, errors, loader_fn)
   if not has_daos then
     return {}
   end
-  local iterator = daos_schemas[1] and ipairs or pairs
+  if not daos_schemas[1] and next(daos_schemas) then
+    -- daos_schemas is a non-empty hash (old syntax). Sort it topologically in order to avoid errors when loading
+    -- relationships before loading entities within the same plugin
+    daos_schemas = sort_daos_schemas_topologically(daos_schemas)
+  end
+
   local res = {}
-  for name, schema_def in iterator(daos_schemas) do
-    if name ~= "tables" and schema_def.name then
-      local ret, err = loader_fn(plugin, schema_def, errors)
-      if err then
-        return nil, err
-      end
-      res[schema_def.name] = ret
+  local schema_def, ret, err
+  for i = 1, #daos_schemas do
+    schema_def = daos_schemas[i]
+    ret, err = loader_fn(plugin, schema_def, errors)
+    if err then
+      return nil, err
     end
+    res[schema_def.name] = ret
   end
 
   return res