+++ kong/plugins/datadog/schema.lua
@@ -23,6 +23,7 @@ local STAT_TYPES = {
   "meter",
   "set",
   "timer",
+  "distribution",
 }
 
 local CONSUMER_IDENTIFIERS = {