+++ spec/01-unit/09-balancer/02-least_connections_spec.lua
@@ -6,7 +6,7 @@
 -- [ END OF LICENSE 0867164ffc95e54f04670b5169c09574bdbd9bba ]
 
 
-local dns_utils = require "resty.dns.utils"
+local dns_utils = require "kong.resty.dns.utils"
 local mocker = require "spec.fixtures.mocker"
 local utils = require "kong.tools.utils"
 
@@ -165,10 +165,10 @@ describe("[least-connections]", function()
   local snapshot
 
   setup(function()
-    _G.package.loaded["resty.dns.client"] = nil -- make sure module is reloaded
+    _G.package.loaded["kong.resty.dns.client"] = nil -- make sure module is reloaded
     _G.package.loaded["kong.runloop.balancer.targets"] = nil -- make sure module is reloaded
 
-    client = require "resty.dns.client"
+    client = require "kong.resty.dns.client"
     targets = require "kong.runloop.balancer.targets"
     balancers = require "kong.runloop.balancer.balancers"
     local healthcheckers = require "kong.runloop.balancer.healthcheckers"